bcm2836_control: add emulation of root interrupt controller on pi2
This adds a first version of emulation for the bcm2836 root interrupt
controller (the new root interrupt controller on raspberry pi 2). This
will ultimately also need to support local timers/mailboxes and
routing of some additional interrupts. This change also implements the
mailbox interrupt control registers (but not the mailboxes
themselves), to avoid needless spam in debug output.