raspi2: refactor CPU init and boot code, and fix a lurking bug
This change moves the code to create the CPU and hook up the interrupt
lines out to a helper function, which should enable multi-core
emulation down the road (although at present the interrupt controller
doesn't support it). This also adds code to set the periphbase/CBAR
value in CP15 -- previously this defaulted to zero, which was causing
UEFI to trash some data in the first page of memory.

I've also cleaned up the boot logic to be more aligned with the other
boards. Now, you load UEFI by passing -bios, rather than the old
kludge that requires both -kernel and -initrd to be the same path.