arm: raise fault for unaligned addresses in LDREX
ARM requires alignment checks for LDREX (and a number of other cases)
which qemu doesn't implement. However, Windows depends on taking an
alignment exception on an LDREX at boot, and won't boot
otherwise. This change implements an alignment check in LDREX, and a
new helper/infrastructure for reporting them. It does not implement
alignment checks for any other instructions.

Since I wasn't sure whether it was a good idea to enable this for
everyone (it will have a small performance impact), I added a #define
CONFIG_ALIGNMENT_EXCEPTIONS, but have not yet hooked that up to any
configure machinery.