sdhci: add Pi-specific quirk/kludge to set card insertion insterrupt at startup
This one is pretty unsavoury in its current format, but is the minimal
change needed to keep UEFI booting with the common sdhci
emulation. The observed behaviour of the Pi2 hardware appears to be
that card insertion interrupts are visible even if they occur when the
interrupt status is initially disabled (i.e., if a card is inserted at
power on, then on a later write to enable the card insertion
interrupt, it still appears immediately). However, the SDHCI spec (sec
1.8) is pretty clear that the interrupt status enable register acts as
a mask before the interrupt state is latched; i.e. if the interrupt
occurs but it is disabled, then it is forever lost. It's not yet clear
to me how to resolve the hardware inconsistency with the spec /
implementation, nor how to implement it in a way that's acceptable
upstream.