sdhci: back out most of noeject_quirk, which broke pi1
This will need to be cleaned up in upstream. The previous
implementation avoided triggering a card insert interrupt on reset,
but that interrupt never comes anyway because the reset clears all the
registers, so the first sdhci_insert_eject_cb never sees the card
insert interrupt enabled. An unintended side-effect which broke Pi1
raspbian was that not running the insert callback resulted in prnsts
set to zero, which was treated as a read-only card.